requirements.txt
Flask==2.2.2
binance-connector==2.0.0
pandas==1.5.3
```